+++ Pod/SyneriseSDK.framework/Headers/SNRContentWidgetImageButtonCustomAction.h
@@ -8,7 +8,7 @@
 
 #import "SNRContentWidgetBaseCustomAction.h"
 
-typedef void (^SNRContentWidgetImageButtonCustomActionReceiveClickActionBlock)(SNRBaseModel * _Nullable model, BOOL isSelected);
+typedef void (^SNRContentWidgetImageButtonCustomActionReceiveClickBlock)(SNRBaseModel * _Nullable model, BOOL isSelected);
 typedef BOOL (^SNRContentWidgetImageButtonCustomActionIsSelectedBlock)(SNRBaseModel * _Nullable model);
 
 NS_ASSUME_NONNULL_BEGIN
@@ -26,7 +26,7 @@ NS_SWIFT_NAME(ContentWidgetImageButtonCustomAction)
 @property (strong, nonatomic, nullable, readwrite) UIImage *selectedImage;
 @property (strong, nonatomic, nullable, readwrite) SNRContentWidgetImageButtonCustomActionIsSelectedBlock isSelected;
 
-@property (strong, nonatomic, nullable, readwrite) SNRContentWidgetImageButtonCustomActionReceiveClickActionBlock onReceiveClickAction;
+@property (strong, nonatomic, nullable, readwrite) SNRContentWidgetImageButtonCustomActionReceiveClickBlock onReceiveClick;
 
 @end