+++ Pod/SyneriseSDK.framework/Headers/SNRPromotions.h
@@ -68,7 +68,7 @@ NS_SWIFT_NAME(Promotions)
                         types:(nullable NSArray<NSNumber *> *)types
                          page:(NSInteger)page
                       success:(nullable void (^)(SNRPromotionResponse *promotionResponse))success
-                      failure:(nullable void (^)(NSError *error))failure NS_SWIFT_NAME(getPromotions(statuses:types:page:success:failure:)) DEPRECATED_MSG_ATTRIBUTE("Use SNRPromotions.getPromotions(apiQuery:success:failure:) instead.");
+                      failure:(nullable void (^)(NSError *error))failure NS_SWIFT_NAME(getPromotions(statuses:types:page:success:failure:)) DEPRECATED_MSG_ATTRIBUTE("Use Promotions.getPromotions(apiQuery:success:failure:) instead.");
 
 /**
  * Gets promotions that are defined for parameters provided.
@@ -87,7 +87,7 @@ NS_SWIFT_NAME(Promotions)
                         limit:(NSInteger)limit
                          page:(NSInteger)page
                       success:(nullable void (^)(SNRPromotionResponse *promotionResponse))success
-                      failure:(nullable void (^)(NSError *error))failure NS_SWIFT_NAME(getPromotions(statuses:types:limit:page:success:failure:)) DEPRECATED_MSG_ATTRIBUTE("Use SNRPromotions.getPromotions(apiQuery:success:failure:) instead.");
+                      failure:(nullable void (^)(NSError *error))failure NS_SWIFT_NAME(getPromotions(statuses:types:limit:page:success:failure:)) DEPRECATED_MSG_ATTRIBUTE("Use Promotions.getPromotions(apiQuery:success:failure:) instead.");
 
 /**
  * Gets promotions that are defined for parameters provided.
@@ -108,7 +108,7 @@ NS_SWIFT_NAME(Promotions)
                              page:(NSInteger)page
                       includeMeta:(BOOL)includeMeta
                           success:(nullable void (^)(SNRPromotionResponse *promotionResponse))success
-                          failure:(nullable void (^)(NSError *error))failure NS_SWIFT_NAME(getPromotions(statuses:types:limit:page:includeMeta:success:failure:)) DEPRECATED_MSG_ATTRIBUTE("Use SNRPromotions.getPromotions(apiQuery:success:failure:) instead.");
+                          failure:(nullable void (^)(NSError *error))failure NS_SWIFT_NAME(getPromotions(statuses:types:limit:page:includeMeta:success:failure:)) DEPRECATED_MSG_ATTRIBUTE("Use Promotions.getPromotions(apiQuery:success:failure:) instead.");
 
 /**
  * Gets promotion that are defined for UUID parameter provided.